+++ crates/cli/src/commands/test.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::process::Stdio;
 
-use smplx_test::TestConfig;
+use smplx_test::{SMPLX_TEST_MARKER, TestConfig};
 
 use super::core::TestFlags;
 use super::error::CommandError;
@@ -9,11 +9,11 @@ use super::error::CommandError;
 pub struct Test {}
 
 impl Test {
-    pub fn run(config: TestConfig, tests: &[String], flags: &TestFlags) -> Result<(), CommandError> {
+    pub fn run(config: TestConfig, filter: String, flags: &TestFlags) -> Result<(), CommandError> {
         let cache_path = Self::get_test_config_cache_name()?;
         config.to_file(&cache_path)?;
 
-        let mut cargo_test_command = Self::build_cargo_test_command(&cache_path, tests, flags);
+        let mut cargo_test_command = Self::build_cargo_test_command(&cache_path, filter, flags);
 
         let output = cargo_test_command.output()?;
 
@@ -33,10 +33,10 @@ impl Test {
         Ok(())
     }
 
-    fn build_cargo_test_command(cache_path: &PathBuf, tests: &[String], flags: &TestFlags) -> std::process::Command {
+    fn build_cargo_test_command(cache_path: &PathBuf, filter: String, flags: &TestFlags) -> std::process::Command {
         let mut cargo_test_command = std::process::Command::new("sh");
 
-        cargo_test_command.args(["-c".to_string(), Self::build_test_command(tests, flags)]);
+        cargo_test_command.args(["-c".to_string(), Self::build_test_command(filter, flags)]);
 
         cargo_test_command
             .env(smplx_test::TEST_ENV_NAME, cache_path)
@@ -47,20 +47,10 @@ impl Test {
         cargo_test_command
     }
 
-    fn build_test_command(tests: &[String], flags: &TestFlags) -> String {
+    fn build_test_command(filter: String, flags: &TestFlags) -> String {
         let mut command_as_arg = String::new();
 
-        if tests.is_empty() {
-            command_as_arg.push_str("cargo test --tests");
-        } else {
-            let mut arg = "cargo test".to_string();
-
-            for test_name in tests {
-                arg.push_str(&format!(" --test {test_name}"));
-            }
-
-            command_as_arg.push_str(&arg);
-        }
+        command_as_arg.push_str(&format!("cargo test {filter}_{SMPLX_TEST_MARKER}"));
 
         let flag_args = Self::build_test_flags(flags);