+++ crates/sdk/src/program/error.rs
@@ -26,4 +26,7 @@ pub enum ProgramError {
 
     #[error("Input index exceeds u32 maximum: {0}")]
     InputIndexOverflow(#[from] std::num::TryFromIntError),
+
+    #[error("Failed to obtain program witness types: {0}")]
+    ProgramGenAbiMeta(String),
 }