+++ examples/basic/tests/basic_test.rs
@@ -2,7 +2,6 @@ use simplex::simplicityhl::elements::{Script, Txid};
 
 use simplex::constants::DUMMY_SIGNATURE;
 use simplex::transaction::{FinalTransaction, PartialInput, ProgramInput, RequiredSignature};
-use simplex::utils::tr_unspendable_key;
 
 use simplex_example::artifacts::p2pk::P2pkProgram;
 use simplex_example::artifacts::p2pk::derived_p2pk::{P2pkArguments, P2pkWitness};
@@ -14,8 +13,8 @@ fn get_p2pk(context: &simplex::TestContext) -> (P2pkProgram, Script) {
         public_key: signer.get_schnorr_public_key().serialize(),
     };
 
-    let p2pk = P2pkProgram::new(tr_unspendable_key(), arguments);
-    let p2pk_script = p2pk.get_program().get_script_pubkey(context.get_network());
+    let p2pk = P2pkProgram::new(arguments);
+    let p2pk_script = p2pk.get_script_pubkey(context.get_network());
 
     (p2pk, p2pk_script)
 }
@@ -25,10 +24,10 @@ fn spend_p2wpkh(context: &simplex::TestContext) -> anyhow::Result<Txid> {
 
     let (_, p2pk_script) = get_p2pk(context);
 
-    let res = signer.send(p2pk_script.clone(), 50)?;
-    println!("Broadcast: {}", res);
+    let txid = signer.send(p2pk_script.clone(), 50)?;
+    println!("Broadcast: {}", txid);
 
-    Ok(res)
+    Ok(txid)
 }
 
 fn spend_p2pk(context: &simplex::TestContext) -> anyhow::Result<Txid> {
@@ -49,28 +48,28 @@ fn spend_p2pk(context: &simplex::TestContext) -> anyhow::Result<Txid> {
 
     ft.add_program_input(
         PartialInput::new(p2pk_utxos[0].clone()),
-        ProgramInput::new(Box::new(p2pk.get_program().clone()), Box::new(witness.clone())),
+        ProgramInput::new(Box::new(p2pk.as_ref().clone()), Box::new(witness.clone())),
         RequiredSignature::Witness("SIGNATURE".to_string()),
     );
 
-    let res = signer.broadcast(&ft)?;
-    println!("Broadcast: {}", res);
+    let txid = signer.broadcast(&ft)?;
+    println!("Broadcast: {}", txid);
 
-    Ok(res)
+    Ok(txid)
 }
 
 #[simplex::test]
 fn basic_test(context: simplex::TestContext) -> anyhow::Result<()> {
     let provider = context.get_default_provider();
 
-    let tx = spend_p2wpkh(&context)?;
+    let txid = spend_p2wpkh(&context)?;
 
-    provider.wait(&tx)?;
+    provider.wait(&txid)?;
     println!("Confirmed");
 
-    let tx = spend_p2pk(&context)?;
+    let txid = spend_p2pk(&context)?;
 
-    provider.wait(&tx)?;
+    provider.wait(&txid)?;
     println!("Confirmed");
 
     Ok(())